Changed HeightMap entry type to double - Necessary for nextGaussian() use

+++ PrintingUtilities.java
@@ -1,6 +1,6 @@
 public class PrintingUtilities {
-    
-    public static void PrintArray(float[][] arr){
+
+    public static void PrintArray(double[][] arr){
         for(int i = 0 ; i < arr.length; i++){
             for(int j = 0 ; j < arr[0].length; j++){
                 System.out.print(String.format("%.4f", arr[i][j]));
